refactor: extract and replace NormalCard with CardFigure
