+++ packages/dnb-eufemia/src/extensions/payment-card/PaymentCard.js
@@ -23,31 +23,14 @@ import {
   createSkeletonClass,
 } from '../../components/skeleton/SkeletonHelper'
 
-import StatusOverlay, { isCardBlocked } from './components/StatusOverlay'
-import CardNumberText, { formatCardNumber } from './components/CardNumber'
-import CardTypeText from './components/CardType'
+import { formatCardNumber } from './components/CardNumber'
+import CardFigure, { cardDataPropTypes } from "./components/CardFigure";
 
 import { ProductType, CardType, BankAxeptType } from './utils/Types'
 import Designs, { defaultDesign } from './utils/CardDesigns'
 import cardProducts from './utils/cardProducts'
-import {
-  ProductLogo,
-  TypeLogo,
-  BankLogo,
-  BankAxeptLogo,
-} from './icons'
 
 export { Designs, ProductType, CardType, BankAxeptType }
-
-const cardDataPropTypes = PropTypes.shape({
-  productCode: PropTypes.string.isRequired,
-  productName: PropTypes.string.isRequired,
-  displayName: PropTypes.string.isRequired,
-  cardDesign: PropTypes.object.isRequired,
-  cardType: PropTypes.object.isRequired,
-  productType: PropTypes.object.isRequired,
-  bankAxept: PropTypes.object.isRequired,
-})
 
 const translationDefaultPropsProps = {
   text_card_number: null,
@@ -172,7 +155,7 @@ export default class PaymentCard extends React.PureComponent {
                 <figcaption className="dnb-sr-only dnb-payment-card__figcaption">
                   {cardData.productName}
                 </figcaption>
-                <NormalCard
+                <CardFigure
                   id={id}
                   skeleton={isTrue(skeleton)}
                   data={cardData}
@@ -208,93 +191,3 @@ const defaultCard = (productCode) => ({
   productType: ProductType.None,
   bankAxept: BankAxeptType.None,
 })
-
-const CardProvidersSeparator = ({ hasBankAxept, hasCardType }) => {
-  return (hasBankAxept && hasCardType) && (<div className="separator" />)
-}
-
-NormalCard.propTypes = {
-  id: PropTypes.string,
-  skeleton: PropTypes.bool,
-  data: cardDataPropTypes.isRequired,
-  cardStatus: PropTypes.string.isRequired,
-  cardNumber: PropTypes.string.isRequired,
-  translations: PropTypes.object.isRequired,
-}
-
-function NormalCard({
-  data,
-  cardStatus,
-  cardNumber,
-  id = null,
-  skeleton = false,
-  translations,
-}) {
-  const cardClasses = classnames(
-    'dnb-payment-card__card',
-    `dnb-payment-card__${data.cardDesign.cardStyle}`,
-    `${isCardBlocked(cardStatus) ? 'dnb-payment-card__card--blocked' : ''}`,
-  );
-
-  return (
-    <div className="dnb-payment-card__card__wrapper">
-      <div
-          id={id}
-          className={cardClasses}
-          {...(data.cardDesign?.backgroundImage
-              ? {
-                style: {
-                  backgroundImage: `url(${data.cardDesign.backgroundImage})`,
-                },
-              }
-              : {})}
-      >
-        <div className="dnb-payment-card__card__content">
-          <div className="dnb-payment-card__card__top">
-            <div className="dnb-payment-card__card__top__left">
-              <BankLogo logoType={data.cardDesign.bankLogo} />
-              <ProductLogo
-                  productType={data.productType}
-                  cardDesign={data.cardDesign}
-              />
-            </div>
-            <div className="dnb-payment-card__card__top__right">
-              <CardTypeText
-                  isCredit={data.bankAxept === BankAxeptType.Credit}
-                  translations={translations}
-                  skeleton={skeleton}
-              />
-            </div>
-          </div>
-          <div className="dnb-payment-card__card__bottom">
-            <div className="dnb-payment-card__card__bottom__left">
-              <CardNumberText
-                  cardNumber={cardNumber}
-                  skeleton={skeleton}
-              />
-            </div>
-            <div className="dnb-payment-card__card__bottom__right">
-              <BankAxeptLogo
-                  bankAxept={data.bankAxept}
-                  cardDesign={data.cardDesign}
-              />
-              <CardProvidersSeparator
-                  hasBankAxept={data.bankAxept === BankAxeptType.BankAxept}
-                  hasCardType={data.cardType !== CardType.None}
-              />
-              <TypeLogo
-                  cardType={data.cardType}
-                  cardDesign={data.cardDesign}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-      <StatusOverlay
-          cardStatus={cardStatus}
-          cardDesign={data.cardDesign.cardStyle}
-          translations={translations}
-      />
-    </div>
-  )
-}
